add Remark entity

+++ src/main/java/ekindergarten/controller/ParentController.java
@@ -9,6 +9,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.math.BigDecimal;
 import java.util.Set;
 
 @RestController
@@ -34,7 +35,7 @@ public class ParentController {
     }
 
     @GetMapping("/getBalance/{childId}")
-    public double getUserBalanceStatusForSpecificChild(@PathVariable long childId) {
+    public BigDecimal getUserBalanceStatusForSpecificChild(@PathVariable long childId) {
         return childService.getSpecificChildBalance(childId);
     }
